Fix to work on Rust alpha nightly on OSX.
Note that the glfw-rs is taken from local repo, because the official current version won't compile due to box syntax issue.

+++ src/buffer/indexbuffer.rs
@@ -38,15 +38,15 @@ impl<'a> IndexBufferEditor<'a> {
         self.data(data);
     }
 
-    pub fn sub_data_u8(&mut self, data: &[u8], byte_offset: uint) {
+    pub fn sub_data_u8(&mut self, data: &[u8], byte_offset: usize) {
         self.sub_data(data, byte_offset);
     }
 
-    pub fn sub_data_u16(&mut self, data: &[u16], byte_offset: uint) {
+    pub fn sub_data_u16(&mut self, data: &[u16], byte_offset: usize) {
         self.sub_data(data, byte_offset);
     }
 
-    pub fn sub_data_u32(&mut self, data: &[u32], byte_offset: uint) {
+    pub fn sub_data_u32(&mut self, data: &[u32], byte_offset: usize) {
         self.sub_data(data, byte_offset);
     }
 
@@ -56,7 +56,7 @@ impl<'a> IndexBufferEditor<'a> {
         }
     }
 
-    fn sub_data<D>(&mut self, data: &[D], byte_offset: uint) {
+    fn sub_data<D>(&mut self, data: &[D], byte_offset: usize) {
         if let Some(ref index_buffer) = self.vertex_array.index_buffer() {
             index_buffer.sub_data(data, byte_offset);
         }
